feat: resolve comments
