+++ kanban-app/src/app/constants.ts
@@ -2,21 +2,6 @@ export const APP_URL = 'https://radiant-earth-37235.herokuapp.com';
 
 export const TOKEN = 'token';
 
-export enum QUERY_PARAMS_FIRST {
-  signin = '/signin',
-  users = '/users',
-  signup = '/signup',
-  boards = '/boards',
-  file = '/file',
-  logs = '/logs',
-}
-
-export enum QUERY_PARAMS_SECOND {
-  columns = '/columns',
-  error = '/error',
-  info = '/info',
-}
-
 export const user2 = {
   name: 'Alex',
   login: 'user002',
