+++ kanban-app/src/app/enums.ts
@@ -1,0 +1,14 @@
+export enum QUERY_PARAMS_FIRST {
+  signin = '/signin',
+  users = '/users',
+  signup = '/signup',
+  boards = '/boards',
+  file = '/file',
+  logs = '/logs',
+}
+
+export enum QUERY_PARAMS_SECOND {
+  columns = '/columns',
+  error = '/error',
+  info = '/info',
+}
