+++ kanban-app/src/app/pages/welcome/welcome.component.html
@@ -1,4 +1,5 @@
 <h1>{{ 'Welcome to' | translate }} {{ title | translate }}!</h1>
+<h1>{{ 'Eng.cat' | translate }}</h1>
 {{ (user$ | async)?.name }}
 
 <button (click)="logUser()">logUser</button>
