+++ kanban-app/src/app/redux/actions/user.actions.ts
@@ -5,10 +5,3 @@ export const loginUser = createAction(
   '[User] Login User',
   props<{ user: GetUserByIdResponse }>()
 );
-
-export const GetUsersStatus = createAction('[User] Get Users Status');
-
-export const loadUsersFailure = createAction(
-  '[User] Load Users Failure',
-  props<{ error: any }>()
-);
