--- kanban-app/src/app/shared/pipe/translate.pipe.ts
@@ -1,13 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-import { TranslateService } from 'src/app/services/translate.service';
-
-@Pipe({
-  name: 'translate',
-  pure: false,
-})
-export class TranslatePipe implements PipeTransform {
-  constructor(private translate: TranslateService) {}
-  transform(key: any): any {
-    return this.translate.data[key] || key;
-  }
-}
