+++ kanban-app/src/assets/i18n/en.json
@@ -1,5 +1,9 @@
 {
   "Welcome to": "Welcome to",
   "HELLO": "hello {{value}}",
-  "TITLE": "My application (en)"
+  "TITLE": "My application (en)",
+  "Eng": {
+    "cat": "meou",
+    "dog": "second"
+  }
 }
