+++ kanban-app/src/assets/i18n/ru.json
@@ -1,5 +1,9 @@
 {
   "Welcome to": "Добро пожаловать в ",
   "HELLO": "Привет {{value}}",
-  "TITLE": "Мое приложение (ru)"
+  "TITLE": "Мое приложение (ru)",
+  "Eng": {
+    "cat": "МЯУ",
+    "dog": "ГАФ"
+  }
 }
